# Conflicts:
#	formidling/src/test/java/no/nav/ung/sak/formidling/BrevScenarioer.java

+++ formidling/src/main/java/no/nav/ung/sak/formidling/template/dto/innvilgelse/beregning/BarnetilleggDto.java
@@ -2,7 +2,7 @@ package no.nav.ung.sak.formidling.template.dto.innvilgelse.beregning;
 
 public record BarnetilleggDto(
     int antall,
-    int dagsats,
+    long dagsats,
     long dagsatsTotal
 ) {
 }
